# swagger.yaml
openapi: 3.0.0
info:
  title: Cirkle API
  version: 1.0.0
  description: |
    ## Cirkle — Social Events Platform API

    Cirkle is an AI-native social discovery platform for people to find their tribe and attend events together.

    ### Authentication
    All endpoints except `GET /cities`, `GET /lifestyle-tags`, and `POST /auth/login` require a Bearer token.

    Include the token in every protected request:
    Authorization: Bearer <your_jwt_token>

    ### Price Format
    All prices are in **paise** (₹1 = 100 paise). Example: ₹500 = `50000`.

    ### Phone Format
    Indian mobile numbers only. Format: `+91XXXXXXXXXX` where the 10-digit number starts with 6, 7, 8, or 9.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: List my tickets
      description: |
        The My Tickets list. Pure proof-of-purchase logistics — never any group, anchor, or payment-flow data, by design.
      parameters:
        - in: query
          name: filter
          schema:
            type: string
            enum: [upcoming, past]
            default: upcoming
          description: Upcoming events (not yet started) or past events.
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/TicketListItem'
        '400':
          description: Invalid filter value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: Ticket detail
      description: |
        Full ticket detail including the booking reference the QR encodes and the price paid. Scoped to the owner. No group information appears here.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    $ref: '#/components/schemas/TicketDetail'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─── VIBES ────────────────────────────────────────────────────────────
...
  /admin/admins:
    post:
      tags: [Admin — Admins]
      summary: Create an admin account
      description: |
        Creates a new admin (business_development or administrative). This is how administrative admins create BD team accounts.

        Requires the manage_admins capability — a business_development admin gets 403.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
      responses:
        '201':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error (email format, password length, missing displayName, invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability (business_development role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: You do not have permission to manage admin accounts
        '409':
          description: An admin with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Admin — Admins]
      summary: List all admins
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of admin accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  admins:
                    type: array
                    items:
                      $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/admins/{id}:
    get:
      tags: [Admin — Admins]
      summary: Get a single admin account
      description: Same shape as the list rows. password_hash is never returned.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Admin — Admins]
      summary: Update an admin account
      description: |
        Partial update — only fields present in the body are changed. Commonly used to deactivate an admin (`isActive: false`), which invalidates their existing token immediately (authenticateAdmin re-checks is_active on every request, not just at token issuance).

        **Lockout protection.** Two rules make admin management unbrickable. Both return 409 with a machine-readable `error` so the client can say why:

        - `cannot_deactivate_self` — nobody may deactivate their own account, unconditionally, however many other admins exist. The failure mode is losing the session you are in.
        - `last_administrative_admin` — the population of ACTIVE administrative admins may never reach zero, whether by deactivating the last one or demoting them to business_development.

        Self-demotion is deliberately NOT blanket-blocked: it is governed by the second rule alone, so an administrative admin may hand over and step down to BD as long as somebody else remains administrative.

        The population count is taken under a row lock in the same transaction as the update, so two admins demoting each other simultaneously cannot both pass and leave zero.

        **Password reset.** Send `password` to reset it — alone, or alongside other fields. A body containing only `password` is valid and no longer returns "No fields to update". The value is bcrypt-hashed, never returned and never logged; the audit row records only that a reset occurred, not the value. A reset changes no counts, so the lockout rules never block it — resetting your own password is always allowed.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
                isActive:
                  type: boolean
                password:
                  type: string
                  minLength: 8
                  format: password
                  description: >-
                    New password. Hashed with bcrypt before storage, never
                    returned in any response and never written to logs. May be
                    the only field in the body.
      responses:
        '200':
          description: Admin updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error, or no fields provided to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Blocked by lockout protection. Branch on `error`:

            - `cannot_deactivate_self` — you tried to deactivate your own account.
            - `last_administrative_admin` — the change would leave zero active administrative admins.
            - `concurrent_admin_change` — a competing admin change collided; retry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [cannot_deactivate_self, last_administrative_admin, concurrent_admin_change]
                  message:
                    type: string
                    description: Human-readable. Branch on `error`, not on this.
              examples:
                self:
                  value:
                    error: cannot_deactivate_self
                    message: You cannot deactivate your own account.
                last:
                  value:
                    error: last_administrative_admin
                    message: There must be at least one active administrative admin.
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/tickets:
    get:
      tags: [Admin — Tickets]
      summary: List tickets
      description: |
        Filtered by eventId alone, this is effectively the venue guest list for that event.

        **PII masking by role.** End-user `phone` and `email` are role-gated. An `administrative` admin holds `view_pii` and receives the real values; a `business_development` admin receives MASKED ones — phone as `+91******210` (country code plus last three digits), email as `j****@gmail.com` (first character plus the domain, which is kept for context).

        Masking is applied server-side, so the real value never enters the response body for a BD admin and cannot be recovered from the network tab. There is no reveal endpoint or query flag.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: eventId
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: checkedIn
          schema:
            type: string
            enum: ['true', 'false']
          required: false
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
          required: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of tickets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminTicket'
        '400':
          description: checkedIn must be true or false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tickets/{id}:
    get:
      tags: [Admin — Tickets]
      summary: Get ticket detail
      description: |
        Everything from the list view plus which order it came from and the price paid.

        **PII masking by role.** End-user `phone` and `email` are role-gated: `administrative` admins hold `view_pii` and see real values, `business_development` admins receive masked ones (`+91******210`, `j****@gmail.com`). Masking is server-side, so the real value never enters the response body.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    allOf:
                      - $ref: '#/components/schemas/AdminTicket'
                      - type: object
                        properties:
                          order:
                            type: object
                            properties:
                              id:
                                type: string
                                format: uuid
                              pricePaidPaise:
                                type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login. Valid for 7 days.
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
      required: false
      description: Max rows to return. Default 50, clamped to 100 even if a larger value is requested.
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message here
    Admin:
      type: object
      description: password_hash is never included in any response.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@cirkle.live
        displayName:
          type: string
          example: Founder Admin
        role:
          type: string
          enum: [administrative, business_development]
          description: Only capability difference is manage_admins — administrative has it, business_development doesn't. Otherwise identical access.
        isActive:
          type: boolean
          description: Kill switch. A deactivated admin's existing token stops working immediately, not just at expiry. Absent from the login response (implied true — login already checked it).
        createdAt:
          type: string
          format: date-time
...
    PaginationEnvelope:
      type: object
      description: Shared shape for every list endpoint. `data` is added per-endpoint via allOf.
      properties:
        total:
          type: integer
          description: Total matching rows, ignoring limit/offset — for "showing X-Y of Z" and page controls
          example: 3847
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 100
...
    AdminTicket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingRef:
          type: string
          example: CRKL-7F3A2C9E
        checkedIn:
          type: boolean
        checkedInAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            startsAt:
              type: string
              format: date-time
        user:
          type: object
          description: PII — phone is unmasked for now.
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
              nullable: true
            phone:
              type: string
...
    TicketListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingRef:
          type: string
          example: CRKL-7F3A2C9E
        checkedIn:
          type: boolean
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            categoryId:
              type: string
            startsAt:
              type: string
              format: date-time
            venueName:
              type: string
            venueAddress:
              type: string
            bannerUrl:
              type: string
              nullable: true
              description: Presigned GET URL, ~1 hour TTL. Never a raw S3 key — the bucket is private. Null if the event has no banner.
    TicketDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingRef:
          type: string
          example: CRKL-7F3A2C9E
          description: The string the QR code encodes
        checkedIn:
          type: boolean
        pricePaid:
          type: integer
          example: 47200
          description: In paise
        bookedAt:
          type: string
          format: date-time
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            categoryId:
              type: string
            startsAt:
              type: string
              format: date-time
            venueName:
              type: string
            venueAddress:
              type: string
            bannerUrl:
              type: string
              nullable: true
              description: Presigned GET URL, ~1 hour TTL. Never a raw S3 key — the bucket is private. Null if the event has no banner.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: No token provided
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...